requirements.txt
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
cloudinary==1.19.1
dj-database-url==0.5.0
Django==2.2.7
django-widget-tweaks==1.4.5
google-api-python-client==1.7.11
google-auth==1.8.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
gunicorn==20.0.4
httplib2==0.14.0
idna==2.8
mock==3.0.5
oauthlib==3.1.0
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.7
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.13.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.7
whitenoise==5.0.1
